Prototype Code
In the prototype, the team opted for a smaller water tank, to make it easier to send the water through the system and because this was available from the chemistry/biology engineering department. Also, the team chose to use plastic pipes instead of copper pipes, because it is cheaper and available at the university. The team chose to only show the functionality in the prototype, so the case is a second hand plastic box. The case is also bigger than it would be in real life.